requirements.txt
selenium==4.1.0
times==0.7
Pillow==8.4.0
random-user-agent==1.0.1
google-cloud-storage==1.43.0
DateTime==4.3
pandas==1.3.5
numpy==1.21.4
SQLAlchemy==1.4.28
urllib3==1.26.7
PyMySQL==1.0.2
mysql-connector-python==8.0.27